feat: graphql api with graphql-module v1 (#1147)
