+++ .eslintrc.js
@@ -20,5 +20,6 @@ module.exports = {
     '@typescript-eslint/no-explicit-any': 0,
     '@typescript-eslint/no-non-null-assertion': 0,
     '@typescript-eslint/no-unused-vars': 'error',
+    '@typescript-eslint/no-namespace': 'off'
   },
 };
