+++ packages/graphql-api/src/modules/core/schema/query.ts
@@ -1,0 +1,9 @@
+import { gql } from 'graphql-modules';
+import { CoreModuleConfig } from '..';
+
+export default (config: CoreModuleConfig) =>
+  gql(`
+    ${config.extendTypeDefs ? 'extend' : ''} type ${config.rootQueryName || 'Query'}  {
+        getUser: User
+    }
+`);
