+++ packages/graphql-api/src/modules/core/schema/schema-def.ts
@@ -1,0 +1,14 @@
+import { gql } from 'graphql-modules';
+import { CoreModuleConfig } from '..';
+
+export default (config: CoreModuleConfig) =>
+  config.withSchemaDefinition
+    ? [
+        gql(`
+          schema {
+            query: ${config.rootQueryName || 'Query'}
+            mutation: ${config.rootMutationName || 'Mutation'}
+          }
+        `),
+      ]
+    : [];
